Fixed flare effect on spores or shockant projectiles
Added a copy of the flare sprite that is shared between spore launcher and shockroach. Either weapon effect could display incorrectly. It is easier to fix it this way than write up a bug report on Zandronum tracker. This should hopefully fix what is believed to be a rendering issue.