Update script_compendium.h
LDFC gets +1 in attack speed and defense, and minus 1 in difficulty, Grondoval Class renamed to Knight from puglist cause he's not a fist fighter. 
Gondoval gets +1 in defense, -1 in range, -2 in attack speed. Parias gets +1 in range. 
Strifeguy loses 1 in range but gains 1 in difficulty.
Caleb and Lo Wang have +2 in damage and -2 in defense. Also those 2 and duke have a -1 in difficulty now.