Added blood decals for Duke
They came directly from Duke Nukem 3D, and while they already look similar to the ones from Redneck Rampage, I thought that it would be worth it to include this. I don't know if Duke3D monsters are affected by this, but it should be an easy change if needed.