Some consistencies for new Corvus stuff
Tome of Power cooldown moved back below the 7 indicator.
Sidhe Spellbook cooldown moved to where the Tome of Power cooldown was back when it was the unique.
Firemace is currently off the equipment bar for now and also occupies 8 on the slot loadout.
Equipment hotkeys adjusted accordingly.
ReGZDoomified the pickup messages.
Added new unique stuff to Strife.