egon and displacer hopefully done.
Moved "recharge" sequence over to an idle state.
Added a fourth firing animation for the Gluon Gun, which is really just a busted animation in the original model and a duplicate as a result.
Green laser effects from the Displacer projectile no longer cause blood to appear.
Displacer altfire no longer has recoil.
Displacer now has a second ammo check after the windup sequence that causes it to jump to its dryfire state when the criteria aren't met, much like trying to use the altfire in OpFor when no teleport exists in a map. Because haha.
Apparently, there was another sound in gauss' source code that was meant to play when overcharged, but just didn't for me for some reason. Added.