squeak and grenade should be ready
Snark sounds should play more accurately.
Snarks and grenades now play the select animation after throwing (can be fired as well as switched from), and are also hidden when out of ammo.
Removed Bright from thrown grenades. I still don't get why that was the case.
Grenade is shown being held when cooked, as per Half-Life. Not noticeable unless playing with high FOV values, but that's how it is.