mostly K43 code touchups
The K43 now comes after the Panzerfaust in the selection order.
SBARINFO code for the K43 has been fully added. An icon without the grenade is also available when not using the grenade mode (manually done with the eraser tool in SLADE).
New weapon icons are now resized accordingly.
The grenade mode now runs off a magazine system, much like it did in ET.
Added grenade fire sound. Not as noticeable (to me, anyway) compared to the Garand in ET, but it is there.
When using the K43, the player model is now shown holding a Mauser rifle for the time being.
Alt ammos for Duke 64 are always displayed in red in the normal/automap HUD, similar to Ranger's DOE ammo. UsesSecondaryAmmo support also added.
Cleared tabs in SBARINFO.
Characters meant to get an ally phone now get an ally phone again instead of just Leonard.