massive touchups for massive update
Added Skulltag equipment hotkeys.
RTCW attack triggers updated.
SBARINFOified the rest of the new RTCW stuff.
Apparently, the rifle grenade magazines started out loaded in NQ? This is reflected here now, but the time at which the magazine updates still matches vanilla ET.
Renamed Allied ST weapon files to separate unique2.dec files for consistency.
RTCW weapons now bob when firing and reloading.
Added missing underwater state for K43/M1 Garand grenade. Whoops!
If a RTCW weapon is completely out of ammo, it will switch to the Luger instead of the knife. The exceptions are the Luger itself and the Colt(s) with pistol ammo enabled, which still switch to the knife.