Added a better method of giving Strife subsequent unique ammo
Ammo is now completely filled in accordance to the current ammo cap. If samsara_noammocap is on, the amount given will instead correspond to the amount that is normally capped at the default amount with the backpack taken into account (a unique that uses bullets will give 400 bullets, for example).
CyberMage's unique now benefits from the subsequent ammo fillup in Strife.