mostly secret message stuff
Secret messages that are logged are now shown in netplay.
Taunt cooldowns are now checked for in the secrets.
Hexen secret message now shows the map name similar to the original game.
Added support for GZDoom's cl_showsecretmessage CVar.
Made the Strife MAP14 unique giving stuff script only work on that game and that map.