probably safe timing for a commit
Merged Painkiller primary attack triggers in a similar manner to Doomguy's slot 1 attack trigger.
Adjusted weapon Spawn states and CHEATNOTWEAPON flags accordingly. The former as a "just in case", and the latter since the core weapon no longer crashes on select.
Added a missing Bright keyword to the laser segment.