rtcw grenade/pineapple can now fire on "reload", q2 trap instantly switches
Grenades/pineapples in RTCW technically did not play the select animation, instead going back to idle, and could thus be refired after the fire animation.
The trap in The Reckoning did not really have a select or deselect animation; this effect has been recreated here. It also is now hidden when there isn't enough ammo to use it (consistency change).