sound stuff (mostly on the RTCW side)
Space Seal usefail sound now alternates between the two sounds, but the grunt sound is still the first sound. The Wolf3D wall bump sound is used for the landing sound as well (it's something in terms of consistency with other characters, at least).
RTCW BJ now has a proper landing sound. The grunt sound is currently the pain sound (fall damage sound in RTCW has a similar sound with the landing sound mixed in); not exact in comparison, but it'll do for now.
Added sound aliases for grenade (both hand-thrown and rifle) explosions and underwater explosions. Those are to be used for an upcoming soundpack.
Added a missing ricochet sound. Also corrected bullet hit metal sound entries.