added doomsphere mugshot graphic (minus health changes), some armor stuff
It is now possible to make armor pickups that ignore samsara_noarmorcap. This was mainly done for Multi Runes' armor regeneration rune (a new actor is provided as well!). Lazy, but it works.
LMS armor now ignores samsara_noarmorcap and max. armor, ensuring armor is always set to the selected amount when your armor is below it.
Commented out possibly the rest of the Eradicator altclass stuff.