Fusion Cutter minor rework
Slightly slower projectile speed (it was slower than blaster rifle but not by much), in exchange again for more damage.

Altfire is more interesting in that in SW;DF it used double ammo compared firing one at a time, as the DPS increase was balanced by ammo consumption. I like this mechanic, so it's back. 180 damage per 15 tics is good, until it's 4 shells per 15 tics.