Kyle Katarn explosive adjustment
Detonators did 45, as did Mortars, so they're adjusted to only do 45 if you face detonator, since that's how much you'd take in the original game.

Mines did 90, and are getting a similar treatment.

Damage numbers are also adjusted to reflect this. Was considering +FORCERADIUSDMG on the Mortar, but decided that would probably be going too far.