Adjustments to the Concussion Rifle
I can't believe I got beaten to the rework to the radius tracer system.

Put more damage on the explosion, since in SW:DF it was completely invariant with regards to distance (unique for this particular weapon), while reducing the tracer damage, since the tracer didn't appear on the shooter of the projectile. 120 damage coop in area, 90 in DM.