Restored the Bloodscourge's special ripper ability
For some reason, MageStaffFX2 rips enemies if you inherit from it, despite it having no ripper flags. This is a special kind of ripper that rips everything except bosses, and for a Slot VII, this is an ability that I believe it should absolutely have.

Other tweaks: Increased Hammer of Retribution damage to help it compete with the other Hexen slot 5s in terms of DPS.

Updated changelog.